+++ examples/api/lib/material/page_transitions_theme/page_transitions_theme.0.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 /// Flutter code sample for [PageTransitionsTheme].
