+++ examples/api/test/material/page_transitions_theme/page_transitions_theme.0_test.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_api_samples/material/page_transitions_theme/page_transitions_theme.0.dart'
     as example;
