+++ packages/flutter/test/material/snack_bar_test.dart
@@ -10,6 +10,7 @@ library;
 import 'dart:async';
 import 'dart:ui';
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
